chore: moving packages out of core
we will be reworking the folder/package structure
