+++ backend/src/main/java/ca/bc/gov/app/m/oracle/legacyclient/controller/LegacyClientController.java
@@ -2,7 +2,7 @@ package ca.bc.gov.app.m.oracle.legacyclient.controller;
 
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 
-import ca.bc.gov.app.core.CoreConstant;
+import ca.bc.gov.app.ApplicationConstant;
 import ca.bc.gov.app.m.oracle.legacyclient.entity.ForestClientEntity;
 import ca.bc.gov.app.m.oracle.legacyclient.service.LegacyClientService;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -17,7 +17,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-@Tag(name = CoreConstant.ORACLE_API_TAG)
+@Tag(name = ApplicationConstant.ORACLE_API_TAG)
 @CrossOrigin(origins = "${frontend.url}")
 @RestController
 @RequestMapping("/app/m/legacyclient")
