+++ backend/src/main/java/ca/bc/gov/app/m/oracle/legacyclient/repository/ClientDoingBusinessAsRepository.java
@@ -1,6 +1,6 @@
 package ca.bc.gov.app.m.oracle.legacyclient.repository;
 
-import ca.bc.gov.app.core.repository.CoreRepository;
+import ca.bc.gov.app.repository.CoreRepository;
 import ca.bc.gov.app.m.oracle.legacyclient.entity.ClientDoingBusinessAsEntity;
 import java.util.List;
 import org.springframework.data.domain.Pageable;
