+++ backend/src/main/java/ca/bc/gov/app/m/oracle/legacyclient/repository/ClientPublicViewRepository.java
@@ -1,6 +1,6 @@
 package ca.bc.gov.app.m.oracle.legacyclient.repository;
 
-import ca.bc.gov.app.core.repository.CoreRepository;
+import ca.bc.gov.app.repository.CoreRepository;
 import ca.bc.gov.app.m.oracle.legacyclient.entity.ClientPublicViewEntity;
 import ca.bc.gov.app.m.postgres.client.entity.ClientTypeCodeEntity;
 import org.springframework.data.domain.Page;
