+++ backend/src/main/java/ca/bc/gov/app/m/postgres/client/controller/ClientController.java
@@ -2,8 +2,8 @@ package ca.bc.gov.app.m.postgres.client.controller;
 
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 
-import ca.bc.gov.app.core.CoreConstant;
-import ca.bc.gov.app.core.dto.CodeDescrDTO;
+import ca.bc.gov.app.ApplicationConstant;
+import ca.bc.gov.app.dto.CodeDescrDTO;
 import ca.bc.gov.app.m.postgres.client.service.ClientService;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import java.util.List;
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-@Tag(name = CoreConstant.POSTGRES_API_TAG)
+@Tag(name = ApplicationConstant.POSTGRES_API_TAG)
 @RestController
 @RequestMapping("/app/m/client")
 @RequiredArgsConstructor
