+++ backend/src/main/java/ca/bc/gov/app/m/postgres/client/entity/SubmissionDetailEntity.java
@@ -1,6 +1,6 @@
 package ca.bc.gov.app.m.postgres.client.entity;
 
-import ca.bc.gov.app.core.CoreConstant;
+import ca.bc.gov.app.ApplicationConstant;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -18,7 +18,7 @@ import lombok.NoArgsConstructor;
 import lombok.With;
 
 @Entity
-@Table(name = "SUBMISSION_DETAIL", schema = CoreConstant.POSTGRES_ATTRIBUTE_SCHEMA)
+@Table(name = "SUBMISSION_DETAIL", schema = ApplicationConstant.POSTGRES_ATTRIBUTE_SCHEMA)
 @Data
 @Builder
 @NoArgsConstructor
@@ -29,7 +29,7 @@ public class SubmissionDetailEntity {
   @Id
   @GeneratedValue(generator = "SEQ_SUBMISSION_DETAIL")
   @SequenceGenerator(name = "SEQ_SUBMISSION_DETAIL",
-      sequenceName = CoreConstant.POSTGRES_ATTRIBUTE_SCHEMA_QUALIFIER +
+      sequenceName = ApplicationConstant.POSTGRES_ATTRIBUTE_SCHEMA_QUALIFIER +
           "SUBMISSION_DETAIL_ID_SEQ",
       allocationSize = 1)
   @Column(name = "SUBMISSION_DETAIL_ID", nullable = false, precision = 10, scale = 0)
