+++ backend/src/main/java/ca/bc/gov/app/m/postgres/client/repository/SubmissionDetailRepository.java
@@ -1,6 +1,6 @@
 package ca.bc.gov.app.m.postgres.client.repository;
 
-import ca.bc.gov.app.core.repository.CoreRepository;
+import ca.bc.gov.app.repository.CoreRepository;
 import ca.bc.gov.app.m.postgres.client.entity.SubmissionDetailEntity;
 import org.springframework.stereotype.Repository;
