+++ backend/src/main/java/ca/bc/gov/app/m/postgres/client/repository/SubmissionRepository.java
@@ -1,6 +1,6 @@
 package ca.bc.gov.app.m.postgres.client.repository;
 
-import ca.bc.gov.app.core.repository.CoreRepository;
+import ca.bc.gov.app.repository.CoreRepository;
 import ca.bc.gov.app.m.postgres.client.entity.SubmissionEntity;
 import org.springframework.stereotype.Repository;
